requirements.txt
asgiref==3.8.1
astunparse==1.6.3
backports.zoneinfo==0.2.1
Django==4.2.20
graphql-core==3.2.6
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
strawberry-graphql==0.255.0
strawberry-graphql-django==0.48.0
typing_extensions==4.13.2
